+++ CHANGELOG.md
@@ -5,8 +5,12 @@
 > [!IMPORTANT]
 > Third party CKEditor plugins are now registered as modules. See [#363](https://github.com/craftcms/ckeditor/pull/363) for details on how to update existing Craft plugins.
 
+- The link modal UI now reflects Craft’s native Link fields. ([#375](https://github.com/craftcms/ckeditor/pull/375))
+- Added the “Advanced Link Fields” setting to CKEditor configs. ([#375](https://github.com/craftcms/ckeditor/pull/375))
 - Added `craft\ckeditor\helpers\CkeditorConfig`. ([#363](https://github.com/craftcms/ckeditor/pull/363))
 - Added `craft\ckeditor\web\assets\BaseCkeditorPackageAsset::$namespace`. ([#363](https://github.com/craftcms/ckeditor/pull/363))
+- Added `craft\ckeditor\CkeConfig::$advancedLinkFields`.
+- Added `craft\ckeditor\CkeConfig::advanceLinkOptions()`.
 
 ## Unreleased